refactor fetch step
