+++ .gitignore
@@ -20,5 +20,4 @@ cdk/outputs.json
 .idea/*
 
 # app
-src/fetch/*.yml
-src/invoke/*.yml
+src/fetch/invoke_all/*.yml
