+++ src/fetch/fetch/catalog.yml
@@ -1,0 +1,36 @@
+MEFEjecucionPresupuestalCovid2020:
+  Name: mef_ejecucionpresupuestalcovid2020
+  URI: https://www.datosabiertos.gob.pe/sites/default/files/pcm_covid_2020.zip
+  Filename: mef_ejecucionpresupuestalcovid2020.zip
+MidisBono600:
+  Name: midis_bono600
+  URI: https://www.datosabiertos.gob.pe/sites/default/files/Bono600_DU_010_2021.zip
+  Filename: midis_bono600.zip
+MidisBonoUniversal:
+  Name: midis_bonouniversal
+  URI: https://www.datosabiertos.gob.pe/sites/default/files/BonoUniversal_DU_098_2020.zip
+  Filename: midis_bonouniversal.zip
+MidisBonoFamiliarUniversal:
+  Name: midis_bonofamiliaruniversal
+  URI: https://www.datosabiertos.gob.pe/sites/default/files/BFU_DU_052_2020_Tramo2.zip
+  Filename: midis_bonofamiliaruniversal.zip
+MinsaVacunacion:
+  Name: minsa_vacunacion
+  URI: https://cloud.minsa.gob.pe/s/ZgXoXqK2KLjRLxD/download
+  Filename: minsa_vacunacion.csv
+  CronExpression: '00 05 * * ? *'
+MinsaFallecidosCovid:
+  Name: minsa_fallecidoscovid
+  URI: https://cloud.minsa.gob.pe/s/Md37cjXmjT9qYSa/download
+  Filename: minsa_fallecidoscovid.csv
+  CronExpression: '00 05 * * ? *'
+MinsaPositivosCovid:
+  Name: minsa_positivoscovid
+  URI: https://cloud.minsa.gob.pe/s/Y8w3wHsEdYQSZRp/download
+  Filename: minsa_positivoscovid.csv
+  CronExpression: '00 05 * * ? *'
+SinadefFallecidos:
+  Name: sinadef_fallecidos
+  URI: https://cloud.minsa.gob.pe/s/nqF2irNbFomCLaa/download
+  Filename: sinadef_fallecidos.csv
+  CronExpression: '00 05 * * ? *'
